Rename FinderApi as Search::Query
FinderApi is too broad a name, the single responsibility
of the class should be to make a search query using the
query builders.

Next I'd like to remove the augment_content_item_with_results
stuff, since modifying a provided hash isn't something that
this class should be doing.
