+++ app/models/search_parameters.rb
@@ -26,7 +26,7 @@ class SearchParameters
   def search_term
     full_term = params[:q]&.strip&.gsub(/\s{2,}/, ' ')
     unless full_term.nil?
-      full_term[0, SearchQueryBuilder::MAX_QUERY_LENGTH]
+      full_term[0, Search::QueryBuilder::MAX_QUERY_LENGTH]
     end
   end
