+++ spec/models/search_parameters_spec.rb
@@ -57,7 +57,7 @@ RSpec.describe SearchParameters do
 
   context "#search_term" do
     it "truncates a too-long search query" do
-      max_length = SearchQueryBuilder::MAX_QUERY_LENGTH
+      max_length = Search::QueryBuilder::MAX_QUERY_LENGTH
       long_query = "a" * max_length
       params = search_params("q" => "#{long_query}1234567890")
       expect(params.rummager_parameters[:q]).to eq(long_query)
